Move Python wrapping code for spacetime to spacetime/pymod/
