+++ cpp/modmesh/python/wrapper/module.cpp
@@ -23,7 +23,7 @@
 #include <modmesh/python/wrapper/module.hpp>
 #include <modmesh/pymod/root_pymod.hpp>
 #include <modmesh/onedim/pymod/onedim_pymod.hpp>
-#include <modmesh/python/wrapper/spacetime/spacetime.hpp>
+#include <modmesh/spacetime/pymod/spacetime_pymod.hpp>
 #ifdef QT_CORE_LIB
 #include <modmesh/view/wrap_view.hpp>
 #endif // QT_CORE_LIB
